% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caa-package.R
\docType{package}
\name{caa}
\alias{caa}
\alias{caa-package}
\title{R Wrapper for the Go dnscaa Library}
\description{
Experimental R wrapper for the Go dnscall library. Builds off of work done by
Romain Francois (\url{https://github.com/rstats-go})
}
\seealso{
Useful links:
\itemize{
  \item \url{https://git.rud.is/hrbrmstr/caa}
  \item Report bugs at \url{https://git.rud.is/hrbrmstr/caa/issues}
}

}
\author{
Bob Rudis (bob@rud.is)
}
\keyword{internal}
