% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caa-dig.R
\name{caa_dig}
\alias{caa_dig}
\title{Retrieve the CAA record values for a domain (if any)}
\usage{
caa_dig(x)
}
\arguments{
\item{x}{domain name}
}
\description{
Retrieve the CAA record values for a domain (if any)
}
\examples{
caa_dig("google.com")
}
