% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_test_data.r
\name{get_default_lw_params}
\alias{get_default_lw_params}
\title{Get default length-weight parameters for testing}
\usage{
get_default_lw_params()
}
\value{
List of length-weight parameters for males, females, and unsexed fish
}
\description{
Get default length-weight parameters for testing
}
\keyword{internal}
